\name{assig}
\alias{assig}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Auxiliary function for computing the minimun entropy discretization}
\description{
 This function is used in the computation of the minimum entropy discretization}
\usage{
assig(x, points, nparti, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ A given vector}
  \item{points}{ The nmuber of points}
  \item{nparti}{ The number of partitions}
  \item{n}{ The number of points}
}
\author{Luis Daza}

\seealso{\code{\link{disc.mentr}}}

\keyword{ math}% at least one, from doc/KEYWORDS
