\name{cv10rpart2}
\alias{cv10rpart2}
\title{ Auxiliary function for sequential feature selection}
\description{
 This function finds the number of instances correctly classified by the
decision tree classifier, rpart, using 10-fold cross validation
and one repetition.
}
\usage{
cv10rpart2(datos)
}
\arguments{
  \item{datos}{ The name of the dataset}
}
\author{Edgar Acuna}
\seealso{\code{\link{crossval}}}

\keyword{ classif}
