\name{disc2}
\alias{disc2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Auxiliary function for performing discretization using equal frequency}
\description{
This function is called by the discretization using equal frecuencies. 
}
\usage{
disc2(x, k)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ a numerical vector}
  \item{k}{ the number of intervals}
}
\author{ Edgar Acuna}
\seealso{\code{\link{disc.ef}}}
\keyword{manip}% at least one, from doc/KEYWORDS
