\name{midpoints}
\alias{midpoints}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Auxiliary function for computing minimun entropy discretization }
\description{
This function carries out the partial computation of the minimum entropy 
discretization }
\usage{
midpoints(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ A numerical vector}
}
\author{ Luis Daza}
\seealso{ \code{\link{disc.mentr}}}

\keyword{ math}
