\name{midpoints1}
\alias{midpoints1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Auxiliary function for computing minimun entropy discretization }
\description{
This function finds out a vector of midpoints of a given vector. In case of a type yield zero. Further
adds zero as lasta entry of the midpoints vector.} 
\usage{
midpoints1(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ A numerical vector}
}
\author{ Luis Daza and Edgar Acuna}
\seealso{ \code{\link{disc.mentr}}}

\keyword{ misc}
