\name{near1}
\alias{near1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Auxiliary function for the reliefcont function}
\description{
This function finds the instance in the data matrix 
that is closest to a given instance x. It is assumed that all the
attributes are continuous.
}
\usage{
near1(x, data)
}
\arguments{
  \item{x}{ A given instance }
  \item{data}{The name of the dataset}
}
\author{Edgar Acuna}
\seealso{\code{\link{near2}},\code{\link{relief}}}

\keyword{math}
