\name{near3}
\alias{near3}
\title{Auxiliary function for the reliefcat function}
\description{
This function finds the instance in the data matrix 
that is closest to a given instance x sing teh Manhattan distance. The
attributes can be either continuous or nominal.
}
\usage{
near3(x, data)
}
\arguments{
  \item{x}{ A given instance}
  \item{data}{ The name of the dataset}
}
\author{Edgar Acuna}
\seealso{ \code{\link{relief}},\code{\link{distancia1}}}
\keyword{math}
