\name{redundancy}
\alias{redundancy}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Finding the unique observations in a dataset along with their fequencies}
\description{
This function finds out the unique instances in a dataset along with their fequencies.}
\usage{
redundancy(data)
}
\arguments{
  \item{data}{ The name of the dataset}
}
\author{Edgar Acuna}
\seealso{\code{\link{clean}}}

\keyword{manip}

