\name{score}
\alias{score}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Score function used in the Bay's algorithm }
\description{
This function finds the score function of an instance. This value will be
used to rank the instances as outliers.
}
\usage{
score(data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{namer of the dataset to be used. }
}
\author{ Caroline Rodriguez}
\seealso{\code{\link{baysout}}}
\keyword{ math}% at least one, from doc/KEYWORDS
