\name{tchisq}
\alias{tchisq}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Auxiliary function for the Chi-Merge discretization}
\description{
This function is required to compute the chi-Merge discretization.
}
\usage{
tchisq(obs)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{obs}{ a vector of observed frequencies}
}
\author{ Jaime Porras}
\seealso{\code{\link{chiMerge}}}

\keyword{math}
