\name{vvalen1}
\alias{vvalen1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Auxiliary function for computing  the Van Valen's homocedasticity test}
\description{
This function is required to perform the Van Valen's homocedasticty test.
}
\usage{
vvalen1(data, classn)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{The name of the dataset to be considered}
  \item{classn}{ The class numnber}
}
\author{Edgar Acuna}
\seealso{ \code{\link{vvalen}}}
\keyword{math}
