\name{disc2}
\alias{disc2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Auxiliary function for performing discretization using equal frequency}
\description{
This function is called by the disc.ef function in the dprep library. 
}
\usage{
disc2(x, k)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ A numerical vector}
  \item{k}{ The number of intervals}
}
\author{ Edgar Acuna}
\seealso{\code{\link{disc.ef}}}
\keyword{manip}

