\name{dprep-package}
\alias{dprep-package}
\alias{dprep}
\docType{package}
\title{
Data Preprocessing for supervised classification
}
\description{
Functions for normalization, treatment of missing values,discretization, outlier detection, feature selection, and visualization}
\details{
\packageIndices{dprep}
Data preprocessing package
}
\author{
Maintainer: Edgar Acuna <edgar.acuna@upr.edu>}
\references{
See website academic.uprm.edu/eacuna/dprep.html
}
\keyword{ package }

