\name{dprep-package}
\alias{dprep-package}
\alias{dprep}
\docType{package}
\title{Data Preprocessing for Supervised Classification
}
\description{
Functions for normalization, treatment of missing values, 
discretization, outlier detection, feature selection, and visualization}
\details{
\tabular{ll}{
Package: \tab dprep\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2008-04-14\cr
License: \tab GPL (version 2 or later) See file LICENCE.\cr
}
For an overview of how to use the package, see user manual provided.
}
\author{
Maintainer: Edgar Acuna <edgar@cs.uprm.edu>
}
\keyword{ package }
