\name{midpoints}
\alias{midpoints}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Auxiliary function for computing the minimun entropy discretization }
\description{
This function makes partial computation required by the ninimum entropy 
discretization. }
\usage{
midpoints(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ a numerical vector}
}
\author{ Luis Daza}
\seealso{ \code{\link{disc.mentr}}}
\keyword{ math}% at least one, from doc/KEYWORDS
