\name{redundancy}
\alias{redundancy}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Find out the unique observations in a dataset along with their fequencies}
\description{
This function finds out the unique instances in a dataset along with their fequencies.}
\usage{
redundancy(data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{ the name of the dataset}
}
\author{Edgar Acuna}
\seealso{\code{\link{clean}}}
\keyword{manip}% at least one, from doc/KEYWORDS
