\name{row.matches}
\alias{row.matches}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Finding rows in a matrix equal to a given vector}
\description{
This function finds instances in a data matrix that are equal to 
a given instance.
}
\usage{
row.matches(y, X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{y}{ A given instance }
  \item{X}{ A given data matrix}
}
\details{ This function was found in the CRAN mailing list. It seems 
to be authored by B. Venables}
\seealso{\code{\link{redundancy}}}

\keyword{misc}

