\name{score}
\alias{score}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Score function used in Bay's algorithm for outlier detection }
\description{
This function finds the score that is used to rank an instance as an outliers.
}
\usage{
score(data)
}
\arguments{
  \item{data}{ The name of the dataset to be used. }
}
\author{ Caroline Rodriguez}
\seealso{\code{\link{baysout}}}
\keyword{ math}
